+++ crypto/keyring/legacy.go
@@ -2,6 +2,7 @@ package keyring
 
 import (
 	"fmt"
+	"io"
 	"strings"
 
 	"github.com/pkg/errors"
@@ -182,3 +183,44 @@ func (kb dbKeybase) ExportPrivKey(name string, decryptPassphrase string,
 func (kb dbKeybase) Close() error { return kb.db.Close() }
 
 func infoKey(name string) []byte { return []byte(fmt.Sprintf("%s.%s", name, infoSuffix)) }
+
+// InfoImporter is implemented by those types that want to provide functions necessary
+// to migrate keys from LegacyKeybase types to Keyring types.
+type InfoImporter interface {
+	// Import imports ASCII-armored private keys.
+	Import(uid string, armor string) error
+}
+
+type keyringMigrator struct {
+	kr keystore
+}
+
+func NewInfoImporter(
+	appName, backend, rootDir string, userInput io.Reader, opts ...AltKeyringOption,
+) (InfoImporter, error) {
+	keyring, err := New(appName, backend, rootDir, userInput, opts...)
+	if err != nil {
+		return keyringMigrator{}, err
+	}
+	kr := keyring.(keystore)
+	return keyringMigrator{kr}, nil
+}
+
+func (m keyringMigrator) Import(uid string, armor string) error {
+	_, err := m.kr.Key(uid)
+	if err == nil {
+		return fmt.Errorf("cannot overwrite key %q", uid)
+	}
+
+	infoBytes, err := crypto.UnarmorInfoBytes(armor)
+	if err != nil {
+		return err
+	}
+
+	info, err := unmarshalInfo(infoBytes)
+	if err != nil {
+		return err
+	}
+
+	return m.kr.writeInfo(uid, info)
+}
